Custom XMP Data is not recognized by Acrobat
iTextSharp is adding an XML declaration as if this were an XML file but it is not an XML file it's a fragment. Adobe Acrobat ignores the XMP data as if it were not valid. This fixes it, and also formats the XML to more closely match a sample file I was given.